sorted services so that they always start in correct order
In older versions of docker-compose (Say 1.23), small numbers of containers get scheduled to start up in the order in which they are listed. In this case, it was trying to start up the api first, which hangs since it cannot open connections to postgres and vault. If we put the api at the bottom of the file it will always start after postgres and vault are up.