create getEnv utility
I wanted to reduce the amount of
boilerplater in terms for extracting
values from environment variables and defaulting to unset ones.

The need came from me wanting to run the program locally and within
the tests container.

Added a simple test case to confirm my logic is ok
and to avoid regressions in the future.

Must investigate how to separate integration and unit tests.

I opted to keep this test in the "env" package since the method is
technically private. As this is an implementation detail I generally err
on not creating tests on internal details but if a bug does arise
I want these tests to fail if the function is the problem.

Long story short, I want to make a tree and find the lowest level in
the  call stack that is failing. That would be the root cause of the
the bug.