implement fetch and add initial test case
Started with a simple test case to check param validation to get a
feel for how to make http requests and handle their responses.

The first instinct was to handle two most common status codes: 200,
and 400. I opted to use a the mapping pattern because I found if else
chains and switch statements to be either hard to read or significantly
more error prone.  By using the init() function of the module
we  can guarantee that the map gets initialized once and stays in memory
for as long as necessary (in this case the duration of the program
running this library)